services/message-api/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
python-jose
